Make UAST v2 the default specification in documentation links.
Rename the UAST v1 and v2 spec files to a common format.  Update the summary to
link to v2 as the default, and update v1 to more strongly indicate that v1 is
being replaced and that v2 should be referred. Update existing spec links to
point to the v2 base.

Updates #236.

Signed-off-by: M. J. Fromberger <michael.j.fromberger@gmail.com>